requirements.txt
sqlalchemy==2.0.16
pydantic==1.10.9
alembic==1.11.1
psycopg2-binary==2.9.6
